feat: move doc test part 1
